Added cmd line parser

+++ src/main/java/ua/edu/chmnu/ki/networks/url/download/ProgressIndicate.java
@@ -2,5 +2,5 @@ package ua.edu.chmnu.ki.networks.url.download;
 
 @FunctionalInterface
 public interface ProgressIndicate {
-    void progress(int count, long total);
+    void progress(long count, long total);
 }
